fix changing from skipInterval to skip track
When skipinterval is registered it over rides the skip forward/back button as there are limited slots.

We need to ensure that the skipinterval listeners are de-registered when we update the remote control settings otherwise we are stuck with the skip interval buttons.

For example, in my app I listen to a pod cast which uses the skip interval control, then I switch to listening to music, the remote controls stay as skip interval regardless of the settings passed to the plugin as once the native listeners are registered they need to be manually unregistered.